Removing edit and summary view and merging to one component
