+++ auth-api/src/auth_api/services/org.py
@@ -387,7 +387,7 @@ class Org:  # pylint: disable=too-many-public-methods
         if org is None:
             raise BusinessException(Error.DATA_NOT_FOUND, None)
 
-        check_auth(token_info, one_of_roles=ADMIN, org_id=org_id)
+        check_auth(token_info, one_of_roles=(ADMIN, STAFF), org_id=org_id)
 
         existing_login_option = AccountLoginOptionsModel.find_active_by_org_id(org_id)
         if existing_login_option is not None:
