+++ auth-web/src/util/constants.ts
@@ -44,7 +44,7 @@ export enum Pages {
     SETUP_ACCOUNT_NON_BCSC_INSTRUCTIONS = 'instructions',
     SETUP_ACCOUNT_NON_BCSC_DOWNLOAD = 'download',
     EDIT_ACCOUNT_TYPE= '/change-account',
-    STAFF_DASHBOARD= 'searchbusiness',
+    STAFF_DASHBOARD= '/searchbusiness',
     STAFF_SETUP_ACCOUNT = 'staff-setup-account',
     CONFIRM_TOKEN = 'confirmtoken',
 }
@@ -112,10 +112,13 @@ export enum AccessType {
 export enum Permission {
     REMOVE_BUSINESS = 'REMOVE_BUSINESS',
     CHANGE_ADDRESS = 'CHANGE_ADDRESS',
+    VIEW_ADDRESS = 'VIEW_ADDRESS',
     CHANGE_ORG_NAME = 'CHANGE_ORG_NAME',
     INVITE_MEMBERS = 'INVITE_MEMBERS',
     CHANGE_ACCOUNT_TYPE = 'CHANGE_ACCOUNT_TYPE',
     CHANGE_ROLE = 'CHANGE_ROLE',
     RESET_PASSWORD = 'RESET_PASSWORD',
-    VIEW_ACCOUNT = 'VIEW_ACCOUNT'
+    VIEW_ACCOUNT = 'VIEW_ACCOUNT',
+    TRANSACTION_HISTORY = 'TRANSACTION_HISTORY',
+    SET_AUTH_OPTIONS = 'SET_AUTH_OPTIONS'
 }
